+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - merchantCategoryType, description, productId and holderId attributes to IssuingPurchase resource
+### Changed
+- senderTaxId parameter to DynamicBrcode resource
 
 ## [0.4.0] - 2023-05-18
 ### Added